Extract Origin, Destination, and Properties from Model.kt
